requirements.txt
fastapi
numpy
pillow
opencv-python
mozjpeg-lossless-optimization
torch